+++ app/_home-posts-wrapper.tsx
@@ -1,9 +1,9 @@
 'use server';
 
+import { GetPosts } from '@/actions/post.actions';
+import { auth } from '@/auth';
 import PostList from '@/components/post-list/post-list';
 import { CreatePost } from '@/components/post/create-post';
-import { GetPosts } from './api/actions/post.actions';
-import { auth } from './api/auth/[...nextauth]/auth';
 
 export default async function HomepagePostsWrapper() {
   const session = await auth();
