+++ app/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,1 @@
-export { GET, POST } from './auth';
+export { GET, POST } from '@/auth';
