+++ components/post-list/post-list.tsx
@@ -1,15 +1,15 @@
 'use client';
-import { GetPosts, GetPostsParams } from '@/app/api/actions/post.actions';
-import { Config } from '@/config/env';
+import { GetPosts, GetPostsParams } from '@/actions/post.actions';
+import { SinglePost } from '@/components/post/single-post';
 import { Post, PostPaginatedResult } from '@/utils/models';
+import { Config } from 'config/env';
 import { useEffect, useState } from 'react';
-import { SinglePost } from '../post/single-post';
 // inspiration from https://medium.com/@ferlat.simon/infinite-scroll-with-nextjs-server-actions-a-simple-guide-76a894824cfd
+import { CreatePost as FirstPost } from '@/components/post/create-post';
+import PostSkeleton from '@/components/skeleton/post-skeleton';
 import { IconButton, RepostIcon } from 'clada-storybook';
 import { useSession } from 'next-auth/react';
 import { useInView } from 'react-intersection-observer';
-import { CreatePost as FirstPost } from '../post/create-post';
-import PostSkeleton from '../skeleton/post-skeleton';
 
 type Props = {
   postsPaginatedResult: PostPaginatedResult | null;
