+++ components/post/post-content.tsx
@@ -1,12 +1,12 @@
 'use client';
+import { InteractionStrip } from '@/shared/interaction/interaction-strip';
+import { UserImage } from '@/shared/user-image';
+import ZoomImage from '@/shared/zoom-image';
+import { UserHeader } from '@/user/user-header';
 import { Post } from '@/utils/models';
 import { replaceTags } from '@/utils/tags';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { InteractionStrip } from '../shared/interaction/interaction-strip';
-import { UserImage } from '../shared/user-image';
-import ZoomImage from '../shared/zoom-image';
-import { UserHeader } from '../user/user-header';
 
 export const PostContent = ({ post, size }: { post: Post; size?: 'large' }) => {
   const textClasses = `mt-s text-black ${size === 'large' && 'text-[20px] tracking-normal leading-relaxed'}`;
@@ -26,7 +26,11 @@ export const PostContent = ({ post, size }: { post: Post; size?: 'large' }) => {
         </Link>
       </div>
       <UserHeader post={post} showTime={true} useLarge={true}></UserHeader>
-      <div className={textClasses} onClick={(e) => e.stopPropagation()}>
+      <div
+        className={textClasses}
+        onClick={(e) => e.stopPropagation()}
+        style={{ wordBreak: 'break-word' }}
+      >
         <p dangerouslySetInnerHTML={{ __html: replaceTags(post.text)! }}></p>
       </div>
       <ZoomImage src={post?.mediaUrl}></ZoomImage>
