+++ components/reply/create-reply.tsx
@@ -1,10 +1,10 @@
 'use client';
 
+import { CreateContent } from '@/components/post/create-content';
+import { UserHeader } from '@/user/user-header';
 import { Post } from '@/utils/models';
 import { useSession } from 'next-auth/react';
 import { useEffect, useState } from 'react';
-import { CreateContent } from '../post/create-content';
-import { UserHeader } from '../user/user-header';
 
 export const CreateReplyComponent = ({ post }: { post: Post }) => {
   const placeholder = 'Und was meinst du dazu?';
