+++ components/user/friends.tsx
@@ -1,6 +1,6 @@
+import { FollowStateProvider } from '@/user/follow-state-context';
+import { UserList } from '@/user/user-list';
 import { useFollowers, useUsers } from '@/utils/hooks/hooks';
-import { FollowStateProvider } from './follow-state-context';
-import { UserList } from './user-list';
 
 export const Friends = ({ userId }: { userId: string }) => {
   const { users } = useUsers();
