+++ components/user/profile.tsx
@@ -1,13 +1,13 @@
 'use client';
 
+import { TimeDiff } from '@/shared/time-diff';
+import { UserImage } from '@/shared/user-image';
+import { Location } from '@/user/location';
+import { UserHandle } from '@/user/user-handle';
+import { getName } from '@/user/user-utils';
 import { User } from '@/utils/models';
 import { EditIcon, SettingsIcon } from 'clada-storybook';
 import NextImage from 'next/image';
-import { TimeDiff } from '../shared/time-diff';
-import { UserImage } from '../shared/user-image';
-import { Location } from './location';
-import { UserHandle } from './user-handle';
-import { getName } from './user-utils';
 
 const fakeULID = '01HGX50H2SXXPE51S60Q7QY29M';
