+++ tailwind.config.ts
@@ -3,10 +3,10 @@ import type { Config } from 'tailwindcss';
 const config: Config = {
   presets: [require('clada-storybook/tailwind.config.ts')],
   content: [
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
-    './node_modules/clada-storybook/**/*.{js,ts,jsx,tsx}',
+    'pages/**/*.{js,ts,jsx,tsx,mdx}',
+    'components/**/*.{js,ts,jsx,tsx,mdx}',
+    'app/**/*.{js,ts,jsx,tsx,mdx}',
+    'node_modules/clada-storybook/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
     extend: {
